Move gunzip into a for loop
The extraction of the first HPA upload has revealed that some of the gzipped
files are corrupted. The current version of the extraction script will stop the
unzipping for each antibody at the first occurrence of an error.
Using a for loop ensures that other TIFF files in an antibody folder keep
getting extracted.